backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pkg-resources==0.0.0
pycparser==2.20
pygame==2.0.0.dev6
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
